Separate implementation from declaration where possible
